requirements.txt
Django==2.1.7
django-registration-redux==2.0
pytz==2018.7
whitenoise==3.3.1
dj-database-url==0.5.0
waitress==1.1.0
idna<2.9,>=2.5
certifi>=2017.4.17
requests
djangorestframework
markdown
django-filter
psycopg2==2.7.3.2
